\name{example}
\alias{example}
\title{Example data}

\description{
A matrix cosists of four columns: The 1st(2nd) column contains the estiamted genetic effects on the outcome (exposure); the 3rd (4th) column contains the estiamted standard errors of the estimated genentic effects on the outcome (exposure)}



