\name{example}
\alias{example}
\title{Example data}

\description{
A matrix consists of four columns: the 1st(2nd) column contains the estimated genetic effects on the outcome (exposure); the 3rd (4th) column contains the estimated standard errors of the estimated genetic effects on the outcome (exposure)}



